// src/app/models/profile.model.ts
export interface IUserInfo {
  _id: string;
  birth_date: string;
  dni: string;
  email: string;
  last_name_father: string;
  last_name_mother: string;
  name: string;
  number_card: string;
  number_document: string;
  phone: string;
  suffix: string;
  type_document: string;
  username: string;
}